Add multi-backend and multi-dtype testing utilities
- Introduced MultiBackendDTypeTest fixture for parameterized testing across different backends and data types.
- Implemented helper functions for dtype-safe tensor operations and model conversion utilities.
- Added backend availability checking with GTEST_SKIP() support.
- Created MultiDTypeTest fixture for testing dtype variations on a single backend.
- Defined macros for instantiating tests with various backend and dtype combinations.

+++ src/backends/cuda/cuda_backend.cpp
@@ -1089,7 +1089,8 @@ public:
                     shape.push_back(std::stoll(shape_str));
                 }
 
-                float value = std::stof(attrs.at("value"));
+                // Use stod for better precision with very small values like denormalized floats
+                float value = static_cast<float>(std::stod(attrs.at("value")));
 
                 DType dtype = DType::Float32;
                 if (attrs.contains("dtype")) {
